requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
cookies==2.2.1
idna==2.6
inflection==0.3.1
Js2Py==0.50
nltk==3.4.5
olefile==0.44
Pillow==6.2.0
postcard-creator==0.0.8
pyjsparser==2.5.2
pypexels==1.0.0b4
python-resize-image==1.1.11
pytz==2017.3
PyYAML==5.3
requests==2.22.0
requests-toolbelt==0.8.0
six==1.11.0
tzlocal==1.4
urllib3==1.25.8
